requirements.txt
# Network scanning and analysis
scapy>=2.4.5
netifaces>=0.11.0
wifi>=0.0.0

# Hardware interfaces
pyric>=0.1.6.3
pywifi>=1.1.12

# Database
sqlalchemy>=1.4.0
alembic>=1.7.5

# UI
prompt_toolkit>=3.0.24
rich>=10.16.1

# Testing
pytest>=6.2.5
pytest-cov>=2.12.1

# Utilities
requests>=2.26.0
pyyaml>=6.0
python-dotenv>=0.19.2 